[레거시 코드 리팩터링 - 3단계] 마루(송채원) 미션 제출합니다. (#378)
* refactor: menu 관련 도메인 jpa entity로 변경

-dao를 repository로 교체

* refactor: MenuGroupService, ProductService에서 dao를 repository로 변경

- 쓰이지 않는 dao 삭제

* refactor: order 관련 도메인을 jpa entity로 변경

- repository 생성

* refactor: Dao, JdbcTemplate 삭제

* refactor: order 관련 service에서 repository 사용하도록 리팩터링

* feat: menu 스냅샷 관리를 위해 OrderHistory 엔티티 추가

- 각 entity 생성자 접근제어자 수정

* test: repository 테스트 추가

* refactor: Order의 OrderTable을 간접 참조로 수정

* refactor: OrderLineItem의 Order를 간접 참조로 수정

* test: 테스트 메서드명 변경

* refactor: JoinColumn 추가

* fix: ungroup tableGroupId exception 발생하도록 수정

* feat: OrderDetail 도메인 추가

- 도메인에서 ungroup 가능 상태를 관리하도록 수정

* refactor: Menu와 MenuProduct 의존 방향 수정

* refactor: TableService에서 OrderRepository 의존 제거

* refactor: Order의 orderTableId 삭제

+++ src/main/java/kitchenpos/ui/dto/MenuProductDto.java
@@ -3,12 +3,12 @@ package kitchenpos.ui.dto;
 public class MenuProductDto {
 
     private Long productId;
-    private Integer quantity;
+    private Long quantity;
 
     private MenuProductDto() {
     }
 
-    public MenuProductDto(Long productId, Integer quantity) {
+    public MenuProductDto(Long productId, Long quantity) {
         this.productId = productId;
         this.quantity = quantity;
     }
@@ -17,7 +17,7 @@ public class MenuProductDto {
         return productId;
     }
 
-    public Integer getQuantity() {
+    public Long getQuantity() {
         return quantity;
     }
 }
